Use public-dns.info information
Turns out that subbrute's list of public DNS servers pales in comparison
to that of https://public-dns.info/. As an example, import the GB valid
list and do a scan of those addresses.

It may be that the importer is unnecessary if the reliability metric of
public-dns.info is accurate.